### 2024-10-17 #7362 @editorialbot
[PRE REVIEW]: IMPPY3D: Image Processing in Python for 3D Image Stacks
**Submitting author:** @NM0ser (<a href="http://orcid.org/0000-0002-3346-6427">Newell Moser</a>)
**Repository:** https://github.com/usnistgov/imppy3d
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
