### 2024-10-17 #7362 @danielskatz
👋 @NM0ser - thanks for your submission.
Note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf`  to make a new PDF. editorialbot commands need to be the first entry in a new comment.
