+++ paper/paper.md
@@ -1,41 +1,39 @@
 ---
-title: >-
-    IMPPY3D: Image Processing in Python for 3D Image Stacks
-authors:
-  - name: Newell H. Moser
-    email: newell.moser@nist.gov
-    affiliation: [1]
-    orcid: 0000-0002-3346-6427
-    corresponding: true
-  - name: Alexander K. Landauer
-    orcid: 0000-0003-2863-039X
-    affiliation: [2]
-  - name: Orion L. Kafka
-    orcid: 0000-0003-2333-8154
-    affiliation: [1]
-affiliations:
-  - index: 1
-    name: Material Measurement Laboratory, National Institute of Standards and Technology, 325 Broadway, Boulder, CO, 80305, USA
-  - index: 2
-    name: Material Measurement Laboratory, National Institute of Standards and Technology, 100 Bureau Drive, Gaithersburg, 20899, MD, USA
-date: 2024-10-17
-bibliography: paper.bib
+title: 'IMPPY3D: Image Processing in Python for 3D Image Stacks'
 tags:
   - python
   - image processing
   - volumetric
   - computed tomography
   - shape characterization
+authors:
+  - name: Newell H. Moser
+    email: newell.moser@nist.gov
+    affiliation: 1
+    orcid: 0000-0002-3346-6427
+    corresponding: true
+  - name: Alexander K. Landauer
+    orcid: 0000-0003-2863-039X
+    affiliation: 2
+  - name: Orion L. Kafka
+    orcid: 0000-0003-2333-8154
+    affiliation: 1
+affiliations:
+  - name: Material Measurement Laboratory, National Institute of Standards and Technology, 325 Broadway, Boulder, CO, 80305, USA
+    index: 1
+  - name: Material Measurement Laboratory, National Institute of Standards and Technology, 100 Bureau Drive, Gaithersburg, 20899, MD, USA
+    index: 2
+date: 17 October 2024
+bibliography: paper.bib
 ---
-
 
 # Summary
 
 Image Processing in Python for 3D image stacks, or IMPPY3D, is a free and open-source software (FOSS) repository that simplifies post-processing and 3D shape characterization for grayscale image stacks, otherwise known as volumetric images, 3D images, or voxel models. While IMPPY3D, pronounced impee-three-dee, was originally created for post-processing image stacks generated from X-ray computed tomography (XCT) measurements, it can be applied generally in post-processing 2D and 3D images. IMPPY3D includes tools for segmenting volumetric images and characterizing the 3D shape of features or regions of interest. These functionalities have proven useful in 3D shape analysis of powder particles, porous polymers, concrete aggregates, internal pores/defects, and more (see the Research Applications section). IMPPY3D consists of a combination of original Python scripts, Cython extensions, and convenience wrappers for popular third-party libraries like SciKit-Image, OpenCV, and PyVista [@scikit-image; @opencv_library; @sullivan2019pyvista].
 
[…]
