### 2024-10-17 #7362 @NM0ser
@danielskatz
Sorry about that! I clicked everything saying I had checked that it compiled -- and it did in the local Docker image -- but I guess there was some formatting issues in the metadata fields at the beginning. Looks like it worked this time.
