### 2024-10-17 #7362 @danielskatz
@NM0ser - please also fix the one DOI mentioned above with the extra prefix in the bibtex doi field.
