+++ paper/paper.bib
@@ -259,7 +259,7 @@
 	volume = {420},
 	issn = {0019-1035},
 	url = {https://www.sciencedirect.com/science/article/pii/S0019103524002264},
-	doi = {https://doi.org/10.1016/j.icarus.2024.116166},
+	doi = {10.1016/j.icarus.2024.116166},
 	journal = {Icarus},
 	author = {Goguen, Jay and Sharits, Andrew and Chiaramonti, Ann and Lafarge, Thomas and Garboczi, Edward},
 	year = {2024},
