### 2024-10-17 #7362 @NM0ser
Great, thanks! I'll work on this list of potential reviewers. We actually have an internal reviewing process for publications before we are allowed to submit a manuscript to a journal, so I may be able to call upon the same internal reviewers.
