### 2024-10-22 #7362 @NM0ser
Here is a list of potential reviewers from [the list of people](https://reviewers.joss.theoj.org/lookup):
jrkerns
sitic
ianhi
jonbmartin
ThomasA

And two more outside of that list:
dwatr001
gknapp1
