### 2024-10-22 #7362 @danielskatz
👋 @jrkerns, @sitic, @ianhi - would one or more of you be willing to review this JOSS submission?
