### 2024-10-23 #7362 @danielskatz
👋 @jonbmartin & @ThomasA - would one of you be willing to review this JOSS submission?
