### 2024-10-23 #7362 @jonbmartin
Hi Daniel,

I am not available to review this article at the moment, as I am currently
reviewing another article for JOSS. But I will be happy to review in the
future.

Jonathan

On Wed, Oct 23, 2024 at 2:33 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @jonbmartin <https://github.com/jonbmartin> & @ThomasA
> <https://github.com/ThomasA> - would one of you be willing to review this
> JOSS submission?
>
