+++ src/imppy3d/ski_processing_wrappers.py
@@ -6,8 +6,8 @@ from skimage import filters as filt
 from skimage.util import img_as_ubyte, img_as_float
 
 # Import local packages
-import ski_processing_wrappers as wrap
-import ski_interactive_processing as ifun
+from . import ski_processing_wrappers as wrap
+from . import ski_interactive_processing as ifun
 
 """
 Currently, no processing wrappers have been implemented for 
