### 2024-11-04 7ba4823
Remove superfluous "Figure" prefix
