### 2024-11-04 2874963
FIx 3DSlicer citation: use official paper
