### 2024-11-04 usnistgov/imppy3d#3 @sitic
[JOSS] Python packages should be `pip install`able
The JOSS review guidelines [state](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions):

> [...] For example, Python packages should be `pip install`able, and have adopted [packaging conventions](https://packaging.python.org/), while Fortran submissions with a Makefile may be sufficient.
>
> **Good:** The software is simple to install, and follows established distribution and dependency management approaches for the language being used
> **OK:** A list of dependencies to install, together with some kind of script to handle their installation (e.g., a Makefile)
> **Bad (not acceptable):** Dependencies are unclear, and/or installation process lacks automation

I think IMPPY3D should be a package that can be imported and should also be `pip install`able. That way it can easily be used in other projects and by all users (e.g. on MacOS, with python versions where the Cython code hasn't been precompiled etc). As anyone with a basic knowledge of python should be able to use IMPPY3D it needs to be very easily installable.

The packaging isn't too complex and shouldn't interfere with the existing Conda/Mamba way too much.

I've created some basic scaffolding to show in principle how to the python packaging. See the PR for details.

A release should then also be uploaded to [PyPi](https://pypi.org), then `pip install imppy3d` should be able to do all the heavy lifting.

Linkback: https://github.com/openjournals/joss-reviews/issues/7405
