### 2024-11-04 #7405 @sitic
JOSS has some strict requirements for the LICENSE file:

> **Acceptable:** A plain-text LICENSE or COPYING file with the contents of an OSI approved license
https://joss.readthedocs.io/en/latest/review_criteria.html#software-license

The current [LICENSE.txt](https://github.com/usnistgov/imppy3d/blob/main/LICENSE.txt)  however contains the MIT text with an additional "NIST Disclaimer" added to the end about government work being the public domain. The two other from NIST which went through JOSS review appear to have been using standard MIT licenses:
* https://github.com/openjournals/joss-reviews/issues/1147
* https://github.com/openjournals/joss-reviews/issues/569

@NM0ser it is possible to change the LICENSE file to be only the MIT part? Perhaps have the NIST Disclaimer somewhere else?
