### 2024-11-06 #7405 @NM0ser
Thank you for your comments and effort! We'll work on the official LICENSE.txt file to be in line with the MIT license exactly, and place the NIST disclaimer elsewhere. The paper proofreading fixes are simple enough as well, thanks for pointing those out.

I'm a bit new to the developer's side of packaging Python, so I appreciate the scaffolding you offered. I'll work to make it pip-installable and once proven successful, publish it to PyPI (this one might take a bit longer as I go through this process for the first time).
