### 2024-11-08 1904a73
Update LICENSE.txt to be exactly MIT license
Added the (unofficial) NIST disclaimer in its own file.
