### 2024-11-12 usnistgov/imppy3d#3 @gknapp1
I haven't tested all of the functionality yet, but a local pip install (`pip install .` from the repo's root directory) does work on the main branch with a minor change. To get working, all version specifiers stating a version "equal to" in setup.py need to use `==` instead of `=`.
