+++ examples/interactive_filtering/interactive_scikit_processing_single_image.py
@@ -12,24 +12,17 @@ using OpenCV.
 """
 
 # Import external dependencies
-import sys
 import numpy as np
-import cv2 as cv
-import skimage.feature as feature
-from skimage import morphology as morph
-from skimage import filters as sfilt
-from skimage.util import img_as_ubyte, img_as_float, img_as_bool
+from skimage.util import img_as_ubyte, img_as_bool
 from skimage.util import invert as ski_invert
 from matplotlib import pyplot as plt
-from matplotlib.widgets import TextBox, Slider, Button, RadioButtons
 
 # Import local modules
-# Ensure this is the correct path to the functions folder
-sys.path.insert(1, '../../functions') 
-import import_export as imex
-import plt_wrappers as pwrap
-import ski_driver_functions as sdrv
-import cv_processing_wrappers as cwrap
+import imppy3d.cv_processing_wrappers as cwrap
+import imppy3d.import_export as imex
+import imppy3d.plt_wrappers as pwrap
+import imppy3d.ski_driver_functions as sdrv
+
 
 # Set constants related to plotting (for MatPlotLib)
 SMALL_SIZE = 10
@@ -88,108 +81,9 @@ print(f"\nInitiating interactive non-local means denoising...")
 img2 = sdrv.interact_driver_denoise(img2, "nl_means")
 
 
-# --------- INTERACTIVE CANNY EDGE --------
-
-def interact_canny_edge(img_in):
-
-    img_0 = img_in.copy()
-
-    # Initial values for the filter
-    sigma_0 = 2.0
-    low_thresh_0 = 0.1*255
-    high_thresh_0 = 0.2*255
-
-    img_mask_0 = feature.canny(img_0, sigma=sigma_0, 
-        low_threshold=low_thresh_0, high_threshold=high_thresh_0)
-
-    img_temp = img_0.copy()
-    img_temp[img_mask_0] = img_0[img_mask_0]/2
-
-    # Global variables (within this function) that will be returned. 
-    # Initialize these to the starting values for now.
-    sigma_out = sigma_0
-    low_thresh_out = low_thresh_0
-    high_thresh_out = high_thresh_0
-    img_out = img_temp
-    mask_out = img_as_ubyte(img_mask_0)
-
-    # Create a figure
-    fig_size = (9,9)
-
-    fig, ax = plt.subplots(1,1)
-    fig.set_size_inches(fig_size[0], fig_size[1])
-    ax.set_aspect('equal')
-
-    # Show the image and save the "matplotlib.image.AxesImage"
-    # object for updating the figure later.
-    img_obj = ax.imshow(img_out, cmap='gray', vmin=0, vmax=255)
-    ax.set_xlabel("X Pixel Number")
-    ax.set_ylabel("Y Pixel Number")
-
-    plt.subplots_adjust(bottom=0.26)
-
-    # Create new axes objects for each button/slider/text widget
-    # 4-tuple of floats rect = [left, bottom, width, height]. A new axes 
-    # is added with dimensions rect in normalized (0, 1) units using 
-    # add_axes on the current figure.
-    sigma_ax = fig.add_axes([0.1, 0.11, 0.15, 0.06])
-    sigma_text_box = TextBox(ax=sigma_ax, label='Sigma  ', 
-        initial=str(sigma_0), textalignment='center')
-
-    low_thresh_ax = fig.add_axes([0.45, 0.11, 0.15, 0.06])
-    low_thresh_text_box = TextBox(ax=low_thresh_ax, label='Low Threshold  ', 
-        initial=str(low_thresh_0), textalignment='center')
-
-    high_thresh_ax = fig.add_axes([0.8, 0.11, 0.15, 0.06])
[…]
