+++ setup.py
@@ -2,6 +2,18 @@ from setuptools import setup, find_packages, Extension
 from Cython.Build import cythonize
 import numpy
 
+# Building the package is done using:
+#   python setup.py bdist_wheel sdist
+#
+# This requires pip, wheel, setuptools, numpy, and Cython, and the
+# appropriate C compiler for you python system. For windows, see the
+# Python documentation:
+# https://wiki.python.org/moin/WindowsCompilers
+#
+# After successfully createing a .whl file, it can be installed
+# locally using:
+#    pip install /path/to/file.whl
+
 extensions = [
     Extension(
         "imppy3d.im3_processing",
@@ -12,9 +24,9 @@ extensions = [
 
 setup(
     name='imppy3d',
-    version='0.1.0',
-    author='Newell Moser',  # FIXME
-    author_email='newell.moser@nist.gov',  # FIXME
+    version='0.1.3',
+    author='Newell Moser',  
+    author_email='newell.moser@nist.gov',  
     description='IMPPY3D: A library for processing 3D image stacks',
     long_description=open('README.md').read(),
     long_description_content_type='text/markdown',
@@ -23,18 +35,18 @@ setup(
     package_dir={'': 'src'},
     include_package_data=True,
     install_requires=[
-        'numpy=1.26.*',
-        'scipy=1.11.*',
-        'matplotlib=3.8.*',
-        'scikit-image=0.20.*',
-        'opencv-python=4.6.*',
-        'vtk=9.2.*',
-        'meshio=5.3.*',
-        'imageio>=2.22',
-        'imageio-ffmpeg>=0.4',
-        'pillow>=9',
-        'tifffile>=2022',
-        'imagecodecs>=2022.7',
+        'numpy == 1.26.*',
+        'scipy == 1.11.*',
+        'matplotlib == 3.8.*',
+        'scikit-image == 0.20.*',
+        'opencv-python == 4.6.*',
+        'vtk == 9.2.*',
+        'meshio == 5.3.*',
+        'imageio >= 2.22',
+        'imageio-ffmpeg >= 0.4',
+        'pillow >= 9',
+        'tifffile >= 2022',
+        'imagecodecs >= 2022.7',
     ],
     ext_modules=cythonize(
         extensions,
