+++ README.md
@@ -1,4 +1,4 @@
-# IMPPY3D (v1.1.0)
+# IMPPY3D
 
 Image Processing in Python for 3D image stacks.
 
@@ -76,13 +76,26 @@ for your operating system and use the `pip` command in your Python environment,
 `pip install /path/to/file.whl`
 
 ### Compiling IMPPY3D
-IMPPY3D is largely Python-based, but there are also C extensions, via Cython, that must be compiled in order to use all of the features of IMPPY3D. A `setup.py` file is provided for easier compilation of IMPPY3D. To compile IMPPY3D, open a terminal with your Python environment active and type,
+IMPPY3D is largely Python-based, but there are also C extensions, via Cython,
+that must be compiled in order to use all of the features of IMPPY3D. A
+`setup.py` file is provided for easier compilation of IMPPY3D. To compile
+IMPPY3D, open a terminal with your Python environment active and type,
 
 `python setup.py bdist_wheel sdist` 
 
-This command will use the setuptools library to compile IMPPY3D into a wheel (.whl). Upon successful completion, the binary .whl file will be located in the folder, "./dist/".
+This command will use the setuptools library to compile IMPPY3D into a wheel
+(.whl). Upon successful completion, the binary .whl file will be located in the
+folder, "./dist/". With the new .whl file created, simply follow the
+instructions above stated about installing IMPPY3D using `pip` and a local
+binary file.
 
-It is important to ensure that your Python environment also has the appropriate environment variables set for your C compiler. If setuptools cannot find your compiler, then compilation of IMPPY3D will fail. Furthermore, the same C-compiler as that which was used to compile the specific Python version you are using should also be used to create the IMPPY3D C extensions. For Windows users, see the documentation, [https://wiki.python.org/moin/WindowsCompilers](https://wiki.python.org/moin/WindowsCompilers).
+It is important to ensure that your Python environment also has the appropriate
+environment variables set for your C compiler. If setuptools cannot find your
+compiler, then compilation of IMPPY3D will fail. Furthermore, the same
+C-compiler as that which was used to compile the specific Python version you
+are using should also be used to create the IMPPY3D C extensions. For Windows
+users, see the documentation, [https://wiki.python.org/moin/WindowsCompilers]
+(https://wiki.python.org/moin/WindowsCompilers).
 
 ## Usage Examples 
 A number of example Python scripts are provided in the "./examples/" folder to
