### 2024-11-13 2c7c66b
Updated obsolete dependencies folder
