### 2024-11-13 c5052e8
Update setuptools packaging
