+++ setup.py
@@ -1,7 +1,3 @@
-from setuptools import setup, find_packages, Extension
-from Cython.Build import cythonize
-import numpy
-
 # Building the package is done using:
 #   python setup.py bdist_wheel sdist
 #
@@ -14,17 +10,22 @@ import numpy
 # locally using:
 #    pip install /path/to/file.whl
 
+import os
+from setuptools import setup, find_packages, Extension
+from Cython.Build import cythonize
+import numpy
+
 extensions = [
     Extension(
-        "imppy3d.im3_processing",
-        ["src/imppy3d/cython/im3_processing.pyx"],
+        name="imppy3d.im3_processing",
+        sources=["src/imppy3d/im3_processing.pyx"],
         include_dirs=[numpy.get_include()]
     )
 ]
 
 setup(
     name='imppy3d',
-    version='1.0.1',
+    version='1.0.6',
     author='Newell Moser',  
     author_email='newell.moser@nist.gov',  
     description='IMPPY3D: A library for processing 3D image stacks',
@@ -34,6 +35,10 @@ setup(
     packages=find_packages(where='src'),
     package_dir={'': 'src'},
     include_package_data=True,
+    setup_requires=[
+        "cython", 
+        "numpy",
+    ],
     install_requires=[
         'numpy == 1.26.*',
         'scipy == 1.11.*',
@@ -41,6 +46,7 @@ setup(
         'scikit-image == 0.20.*',
         'opencv-python == 4.6.*',
         'vtk == 9.2.*',
+        'pyvista == 0.34.*',
         'meshio == 5.3.*',
         'imageio >= 2.22',
         'imageio-ffmpeg >= 0.4',
