+++ examples/segment_3d_particles/binary_imgs/ReadMe.txt
@@ -1,0 +1,1 @@
+This folder contains the output binary images.
