### 2024-11-14 d616806
Updated version requirements in setup.py
