+++ setup.py
@@ -25,7 +25,7 @@ extensions = [
 
 setup(
     name='imppy3d',
-    version='1.0.6',
+    version='1.0.8',
     author='Newell Moser',  
     author_email='newell.moser@nist.gov',  
     description='IMPPY3D: A library for processing 3D image stacks',
@@ -40,19 +40,19 @@ setup(
         "numpy",
     ],
     install_requires=[
-        'numpy == 1.26.*',
-        'scipy == 1.11.*',
-        'matplotlib == 3.8.*',
-        'scikit-image == 0.20.*',
-        'opencv-python == 4.6.*',
-        'vtk == 9.2.*',
-        'pyvista == 0.34.*',
-        'meshio == 5.3.*',
-        'imageio >= 2.22',
-        'imageio-ffmpeg >= 0.4',
-        'pillow >= 9',
-        'tifffile >= 2022',
-        'imagecodecs >= 2022.7',
+        'numpy==1.26.*',
+        'scipy==1.11.*',
+        'matplotlib==3.8.*',
+        'scikit-image==0.20.*',
+        'opencv-python==4.6.*',
+        "vtk>9.1,<=9.3",
+        'pyvista==0.34.*',
+        'meshio==5.3.*',
+#        'imageio>=2.22',
+#        'imageio-ffmpeg>=0.4',
+#        'pillow>=9',
+#        'tifffile>=2022',
+#        'imagecodecs>=2022.7',
     ],
     ext_modules=cythonize(
         extensions,
