### 2024-11-18 9570a2b
Update readme.md for version 1.1.2
