### 2024-11-21 usnistgov/imppy3d#3 @gknapp1
@NM0ser, thanks for getting this up on PyPi. I was able to install using Python 3.10.11.

However, to run the examples (specifically the "calc_metrics_pore" example) I needed to additionally pip install the `imagecodecs` package. Once I did that, all the examples ran as expected. I see that `imagecodecs` is commented out in setup.py. Is there a reason for not including that in the build?
