### 2024-11-21 usnistgov/imppy3d#6 @gknapp1
[JOSS] Automated/manual testing for correctness
The example scripts are helpful to test that things are running, but do not help to test the correctness of the installation per the [JOSS testing guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#tests).

Can you include example results for one or more of the examples and a validation function that validates that the user's execution of the example problem produces results that match the expected output? I think that it is fine for this to be manual, for example, there is documentation for how a user should run a validation script. Though I would encourage also looking into solutions like [pytest](https://docs.pytest.org/en/stable/) and [GitHub continuous integration](https://docs.github.com/en/actions/about-github-actions/about-continuous-integration-with-github-actions) to help with maintaining the repository as development continues.

Linkback: https://github.com/openjournals/joss-reviews/issues/7405
