### 2024-11-21 usnistgov/imppy3d#7 @gknapp1
[JOSS] Missing note on community guidelines for contribution
While you have community guidelines on how to get support for issues, there are no guidelines for how third parties can contribute to this project (see [JOSS Review Criteria: Community Guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)). Please address this in your readme.

Linkback: https://github.com/openjournals/joss-reviews/issues/7405
