### 2024-11-21 usnistgov/imppy3d#3 @NM0ser
Thanks so much for the testing, @gknapp1! PyPi has been a learning process, but it looks like it is coming together now. Regarding `imagecodecs`, I had to do some digging to figure out what is going on there. I commented out some library requirements in the `setup.py` file because I wanted to clean things up. I had libraries marked as required dependencies that I do not actually make calls to, though, they are indirectly required because I do, for example, call the scikit-image library which depends on some of those libraries. However, I should not of course require those dependencies then, scikit-image should require those dependencies.

What you found seems to be an interesting exception. As it turns out, scikit-image wraps other image libraries for importing/exporting, and the user (me) can ask to make special use of these other libraries during a function call. So, in order to make it possible to import/export multi-page TIFF files, I specified that `plugin=tifffile` in one of their importers/exporters. Well, as it turns out, the `tifffile` library has optional dependencies if you want extended features. Specifically, for compressing images, you must have `imagecodecs` installed -- however, it is not specified as required by the package as far as I can tell.

Long story short, I do not make any calls directly to `imagecodecs`, but it seems to have sneaked through the cracks. I have included it back into the `setup.py` file. What made this even more difficult to catch on my end is that I apparently have `imagecodecs` cached in my python development environment. So, even after creating a new Python environment, this somehow did not pop up on my end.

Thanks again! The latest version fixes this requirement in `setup.py`.
