### 2024-11-22 56dd1e9
Update analyze_pores.py
