+++ examples/calc_metrics_pores/analyze_pores.py
@@ -63,24 +63,39 @@ on the header strings:
     algorithm. Note, this field value will read “NaN” if the pore 
     contains fewer than 64 voxels. 
 
-    7) “Centroid Img Num. (1-based indexing)”: The coordinates of the 
+    7) “Centroid Img Num. (0-based indexing)”: The coordinates of the 
     centroid can be described using indices representing the image 
     number, row number, and column number. This specific data field 
-    represents the image index of the centroid. Note, 1-based indexing
-    is used here, implying that the first image corresponds to one, not
-    zero. 
+    represents the image index of the centroid. Note, 0-based indexing
+    is used here, implying that the first image corresponds to zero, not
+    one. 
 
-    8) “Centroid Row Num. (1-based indexing)”: The row index of the 
-    centroid of the pore, using 1-based indexing. For a given image, the
+    8) “Centroid Row Num. (0-based indexing)”: The row index of the 
+    centroid of the pore, using 0-based indexing. For a given image, the
     first row starts at the top of the image. Also, see the notes above
     for bullet 7.
 
-    9) “Centroid Col Num. (1-based indexing)”: The column index of the
-    centroid of the pore, using 1-based indexing. For a given image, the
+    9) “Centroid Col Num. (0-based indexing)”: The column index of the
+    centroid of the pore, using 0-based indexing. For a given image, the
     first column starts on the left-side of the image. Also, see the
     notes above for bullet 7.
 
-    10) “BBox Center Img Num. (1-based indexing)”: A rotated bounding
+    10. “Centroid VTK X [um]”: When visualizing the 3D voids in ParaView
+    using the VTK-voxel files, there is a need to convert the image
+    indices into XYZ-coordinates. For complete transparency, the
+    XYZ-coordinates of the centroids of the pores in VTK models
+    are given in microns. This column specifically provides the
+    X-coordinate (in microns) of the given centroid in the VTK model.
+
+    11. “Centroid VTK Y [um]”: The Y-coordinate (in microns) of the
+    given centroid in the VTK model. Also, see notes above for bullet
+    10.
+
+    12. “Centroid VTK Z [um]”: The Z-coordinate (in microns) of the
+    given centroid in the VTK model. Also, see notes above for bullet
+    10.
+
+    13) “BBox Center Img Num. (0-based indexing)”: A rotated bounding
     box was fitted to pores containing at least 64 voxels. The three
     lengths are denoted based on their lengths, L≥W≥T. Specifically, L
     is defined as the longest distance between (the centers of) any two
@@ -88,35 +103,35 @@ on the header strings:
     voxels while remaining orthogonal to L. T is the longest distance
     between any to voxels while remaining orthogonal to L and W. This
     specific data field represents the image index of the center of the
-    bounding box. Note, 1-based indexing is used here, implying that the
-    first image corresponds to one, not zero. Will be “NaN” if the pore
+    bounding box. Note, 0-based indexing is used here, implying that the
+    first image corresponds to zero, not one. Will be “NaN” if the pore
     contains fewer than 64 voxels.
 
-    11) “BBox Center Row Num. (1-based indexing)”: The row index of the
-    center of the bounding box, using 1-based indexing. For a given
+    14) “BBox Center Row Num. (0-based indexing)”: The row index of the
+    center of the bounding box, using 0-based indexing. For a given
[…]
