+++ examples/calc_metrics_pores/labeled_pores/ReadMe.txt
@@ -1,1 +1,0 @@
-This folder contains the labeled pores as images.
