### 2024-11-22 e767a75
Update analyze_pores.py
