### 2024-11-24 #7405 @gknapp1
> @sitic & @gknapp1 - Are you able to make more progress in your reviews at this point? Or are you stuck waiting for changes from @NM0ser?

I'm waiting on some changes to address automated testing and community guidelines, but they should be relatively minor. The installation comments have been addressed.

@danielskatz, one question about data availability. Do the authors need to provide data to exactly reproduce the example figure in their paper? They have examples for individual parts of the analysis, but it doesn't appear they have a singular example to generate the exact analysis in the paper. Feels okay to me as-is, but I'm not clear on JOSS's stance on data for examples like that.
