### 2024-11-24 #7405 @danielskatz
@gknapp1 - JOSS focuses on the software, not reproducibility of the paper.  It's important that the software can be installed and checked.

However, if you think it's essential that this figure be reproducible, you should say so.
