### 2024-12-03 7ae17af
Update readme with a picture
