+++ README.md
@@ -65,10 +65,10 @@ details for each of these installation cases.
 The simplest method of installing IMPPY3D is through [PyPi](https://pypi.org/project/imppy3d/). 
 Installing IMPPY3D from PyPi can be achieved using `pip` via the following,
 
-`pip install imppy3d=1.1.4`
+`pip install imppy3d=1.1.5`
 
 It is important that you explicitly specify the latest version of IMPPY3D, in
-this case, version 1.1.4. Moreover, the pip installation process of IMPPY3D is
+this case, version 1.1.5. Moreover, the pip installation process of IMPPY3D is
 currently restricted to Python 3.10 environments.
 
 ### Installing Using Pip with Local Binary Files
@@ -143,7 +143,9 @@ Y-axis points downward (across the rows). However, the Z-axis in an image stack
 must be carefully considered. In IMPPY3D, we assume that the first image in the
 image stack corresponds to the TOP of the physical volume; every subsequent
 image represents a slice of the volume beneath the first image. In doing so, a
-right-handed coordinate system is naturally created based on the conversion of the indicial coordinates, as stated in the bullet points above.
+right-handed coordinate system is naturally created based on the conversion of the indicial coordinates, as shown in the picture below.
+
+![](./readme_pics/IMPPY3D_Coordinate_System.png)
 
 Tomographic reconstructions from X-ray radiographs often result in the first
 image in the image stack representing the bottom of the physical object. In
@@ -151,7 +153,7 @@ this case, we recommend reversing the image stack prior to any post-processing
 in IMPPY3D. If the image stack is not reversed, then the resultant 3D visuals
 (as .vtk files) will be reflections of the physical object. Reversing the image
 stack can be automatically performed for you by using the 'flipz' parameter
-when importing an image stack using imppy3d.import_export.load_image_seq(). 
+when importing an image stack using `imppy3d.import_export.load_image_seq()`. 
 
 ## Roadmap
 * Convert the comment blocks in function definitions to a common standard for 
