### 2024-12-04 usnistgov/imppy3d#6 @NM0ser
Thanks @gknapp1, I've taken a look at pytest and related frameworks. Thanks for the reference, I see a lot of libraries use this methodology, such as Numpy. Of course, it would have been better to have started with unit testing in mind from the beginning. I did a lot of my own tests for each addition of the code, but I now see the power of automating these tests and keeping them around. It's nice to know that modifications or additions to the code do not lead to unexpected errors in other parts of the code.

However, as you hinted, I think a suite of manual tests for IMPPY3D will be sufficient for the time being. I'll mark this under lessons learned for development of open source code bases. For IMPPY3D, I have added a new folder containing a [test suite](https://github.com/usnistgov/imppy3d/tree/main/tests) of Python scripts. In many of these tests, I create a dummy/reference data set, run some IMPPY3D functions, and then compare the results against the reference data for correctness. At the moment, I am checking the core functionality of IMPPY3D, but there may be edge cases that slipped through. I can add more unit tests in the future as they pop up.

Also, there are instances where I am simply checking to see if the function can be called without throwing an exception. For example, I have written a number of interactive Python wrappers for the scikit-image library. These scripts create an interactive user-interface via matplotlib, which lets the user more easily pick post-processing parameters (e.g., threshold values, denoising values, sharpening values). Since it is an interactive window, there really is not a "correct" output. Scikit-Image already checks their own functions for correctness, so I do not see the need to check that the images are filtering correctly. Moreover, automated unit testing of a user interface seems like a tricky problem. So, for now, I simply call the interactive scripts (without opening a window that halts the program), and ensure that it at least does not throw an error.
