### 2024-12-04 69bf620
Update test_vtk_models.py
