+++ tests/test_vtk_models.py
@@ -42,7 +42,7 @@ def del_all_vtk_files(dir_path='./'):
         file_to_rem.unlink()
 
 
-def make_dummy_img_stack():
+def make_dummy_ellipsoid_img_stack():
     # Draw an ellipsoid
     # ********** USER INPUTS ********** 
     # x^2/a^2 + y^2/b^2 + z^2/c^2 = 1
@@ -58,12 +58,35 @@ def make_dummy_img_stack():
     return img_ellip
 
 
+def make_dummy_box_img_stack(vec_dir='z'):
+    imgs = np.zeros((129,129,129), dtype=np.uint8)
+
+    rad_maj = 31
+    rad_min = 31
+    mid_i = int(np.floor(129/2.0))
+    
+    if vec_dir=='z': # Along image coordinates
+        imgs[mid_i-rad_maj:mid_i+rad_maj+1, mid_i-rad_min:mid_i+rad_min+1,
+            mid_i-rad_min:mid_i+rad_min+1] = 255
+
+    elif vec_dir=='y': # Along row coordinates
+        imgs[mid_i-rad_min:mid_i+rad_min+1, mid_i-rad_maj:mid_i+rad_maj+1, 
+            mid_i-rad_min:mid_i+rad_min+1] = 255
+
+    elif vec_dir=='x': # Along column coordinates
+        imgs[mid_i-rad_min:mid_i+rad_min+1, mid_i-rad_min:mid_i+rad_min+1,
+            mid_i-rad_maj:mid_i+rad_maj+1] = 255
+
+    return imgs
+
+
 def create_vtk_polydata_surf():
     # Creates a surface from a voxel model of an ellipsoid
     test_name = "\nTEST: CREATE VTK POLYDATA SURFACE USING MARCHING CUBES..."
     print(test_name)
 
-    imgs_b = make_dummy_img_stack()
+    imgs_b = make_dummy_box_img_stack()
+    n_voxels = 63*63*63
 
     try:
         verts, faces, normals, vals = vapi.convert_voxels_to_surface(imgs_b)
@@ -72,9 +95,17 @@ def create_vtk_polydata_surf():
         return [1, test_name + " ERROR"] # (True) Errors occurred
 
     try:
-        vtk_obj = vapi.make_vtk_surf_mesh(verts, faces, vals)
+        vtk_obj = vapi.make_vtk_surf_mesh(verts, faces, vals, smth_iter=0)
     except:
         print(f"\nERROR: Failed to create a VTK polydata object.")
+        return [1, test_name + " ERROR"] # (True) Errors occurred
+
+    TOL = 0.05*n_voxels
+    vol_diff = np.absolute(vtk_obj.volume - n_voxels)
+
+    if vol_diff >= TOL:
+        print(f"\nERROR: Created surface model does not contain the correct"\
+            " volume.")
         return [1, test_name + " ERROR"] # (True) Errors occurred
 
     return [0, test_name + " SUCCESS"] # No errors
@@ -85,12 +116,19 @@ def create_vtk_unstructured_grid():
     test_name = "\nTEST: CREATE VTK UNSTRUCTURED VOXEL MODEL (C EXTENSION)..."
     print(test_name)
 
-    imgs_b = make_dummy_img_stack()
+    imgs_b = make_dummy_box_img_stack() # 63 x 63 x 63 cuboid
+    n_voxels = 63*63*63
 
     try:
         vtk_obj = vapi.make_vtk_unstruct_grid(imgs_b, quiet_in=True)
     except:
         print(f"\nERROR: Failed to create VTK voxel model.")
+        return [1, test_name + " ERROR"] # (True) Errors occurred
+
+    # Check that the correct number of voxels were created
+    if vtk_obj.n_cells != n_voxels:
[…]
