### 2024-12-05 65dd980
Update ReadMe.txt for community guidelines
