### 2024-12-05 4fa61cb
Update ReadMe file about community guidelines
