### 2024-12-05 usnistgov/imppy3d#7 @NM0ser
Thanks, I think we have these guidelines clearer now in the updated ReadMe, but let me know if we should add more detail.
