### 2025-01-14 #7405 @jonbmartin
Hi Daniel, I am still conducting review for the MIRTorch submission to JOSS
https://github.com/openjournals/joss-reviews/issues/734, so I am not
available at the moment. I am sorry - I will let you  know if that wraps up
soon.

Jonathan

On Tue, Jan 14, 2025 at 9:57 AM Daniel S. Katz ***@***.***>
wrote:

> 👋 @jonbmartin <https://github.com/jonbmartin> - You had previously
> volunteered to review this, but you were on another review at the time.
> However, we've now lost a reviewer and need another. Would it still be
> possible for you to review this submission?
>
