### 2025-01-17 #7405 @danielskatz
👋 @NM0ser - It looks like the review is complete. The next step is for me to proofread the paper, then I will give you some actions related to potential changes, if needed, and certainly a new release of the software and an archive of it.
