### 2025-01-17 #7405 @danielskatz
@NM0ser - In the second paragraph of the Statement of Need section, there's a list of examples followed by a bunch of references.  Can you put each reference with the corresponding example? And do the same for the pair of examples later in this paragraph.

Can you also remove forced capitalization in the references (except where it is needed for proper names, etc.) The references that appear to have this are: Ferguson, Groeber, Kafka, Moser (the first one), and Stamati.
