+++ paper/paper.md
@@ -38,9 +38,9 @@ Highlighted capabilities of IMPPY3D include: varying image processing parameters
 
 # Statement of Need
 
-Volumetric images commonly arise from nondestructive measurement techniques such as XCT, optical coherence tomography (OCT) or confocal microscopy, or from destructive techniques such as serial sectioning. Volumetric images typically analyzed in IMPPY3D are grayscale representations of a 3D volume of material and contain both internal and external shape information. For example, XCT is commonly used in metal-based additive manufacturing to prevent parts from entering service that contain critical internal defects. Post-reconstruction image analysis software is often employed to post-process volumetric images, such as Dragonfly[^1] and Avizo [@noauthor_dragonfly_2024; @noauthor_avizo_2024]. While closed-source software packages are highly sophisticated tools, they are also inherently limited since users are restricted by the closed-source publishing model. Users may require specific features that are unavailable, and closed-source models can be difficult or impossible to validate and verify. 
+Volumetric images commonly arise from nondestructive measurement techniques such as XCT, optical coherence tomography (OCT) or confocal microscopy, or from destructive techniques such as serial sectioning. Volumetric images typically analyzed in IMPPY3D are grayscale representations of a 3D volume of material and contain both internal and external shape information. For example, XCT is commonly used in metal-based additive manufacturing to prevent parts from entering service that contain critical internal defects. Post-reconstruction image analysis software is often employed to post-process volumetric images, such as Dragonfly[^1] [@noauthor_dragonfly_2024] and Avizo [@noauthor_avizo_2024]. While closed-source software packages are highly sophisticated tools, they are also inherently limited since users are restricted by the closed-source publishing model. Users may require specific features that are unavailable, and closed-source models can be difficult or impossible to validate and verify. 
 
-Non-commercial software packages are also available that post-process volumetric images with varying degrees of generality and openness. While not an exhaustive list, examples include ImageJ/FIJI, 3D Slicer, DREAM.3D, SPAM, and PuMA [@schindelin_fiji_2012; @3DSlicer; @groeber_dream3d_2014; @stamati_spam_2020; @ferguson_update_2021]. However, existing software can be difficult to extend for custom analyses, and/or their current features and strengths lie outside of volumetric segmentation and 3D shape characterization. There are also FOSS packages that specialize in tomographic reconstruction, such as TomoPy and Tomviz [@gursoy_tomopy_2014; @noauthor_tomviz_2024]. However, the focus of IMPPY3D is the segmentation and feature analysis of already-reconstructed 3D image stacks, rather than image reconstruction itself. IMPPY3D is written in straightforward Python that contains internal documentation with the goal of making the library flexible and extensible to anyone with basic knowledge of Python and image processing. The library has been designed to work within an Conda/Miniforge environment for either Windows or Linux machines.  
[…]
