+++ paper/paper.bib
@@ -81,7 +81,7 @@
 }
 
 @article{gursoy_tomopy_2014,
-	title = {{TomoPy}: a framework for the analysis of synchrotron tomographic data},
+	title = {{TomoPy}: {A} framework for the analysis of synchrotron tomographic data},
 	volume = {21},
 	issn = {1600-5775},
 	shorttitle = {{TomoPy}},
@@ -105,7 +105,7 @@
 
 
 @article{schindelin_fiji_2012,
-	title = {Fiji: an open-source platform for biological-image analysis},
+	title = {{FIJI}: {An} open-source platform for biological-image analysis},
 	volume = {9},
 	copyright = {http://www.springer.com/tdm},
 	issn = {1548-7091, 1548-7105},
@@ -136,7 +136,7 @@
 }
 
 @article{groeber_dream3d_2014,
-	title = {{DREAM}.{3D}: {A} {Digital} {Representation} {Environment} for the {Analysis} of {Microstructure} in {3D}},
+	title = {{DREAM}.{3D}: {A} Digital Representation Environment for the Analysis of Microstructure in {3D}},
 	volume = {3},
 	issn = {2193-9764, 2193-9772},
 	shorttitle = {{DREAM}.{3D}},
@@ -153,7 +153,7 @@
 }
 
 @article{stamati_spam_2020,
-	title = {spam: {Software} for {Practical} {Analysis} of {Materials}},
+	title = {{SPAM}: {Software} for Practical Analysis of Materials},
 	volume = {5},
 	copyright = {http://creativecommons.org/licenses/by/4.0/},
 	issn = {2475-9066},
@@ -170,7 +170,7 @@
 }
 
 @article{ferguson_update_2021,
-	title = {Update 3.0 to “{PuMA}: {The} {Porous} {Microstructure} {Analysis} software”, ({PII}:{S2352711018300281})},
+	title = {Update 3.0 to “{PuMA}: {The} Porous Microstructure Analysis software”, ({PII}:{S2352711018300281})},
 	volume = {15},
 	issn = {23527110},
 	shorttitle = {Update 3.0 to “{PuMA}},
@@ -186,7 +186,7 @@
 }
 
 @article{kafka_insitu_2024,
-	title = {A {Technique} for {In}-{Situ} {Displacement} and {Strain} {Measurement} with {Laboratory}-{Scale} {X}-{Ray} {Computed} {Tomography}},
+	title = {A Technique for In-Situ Displacement and Strain Measurement with Laboratory-Scale X-Ray Computed Tomography},
 	issn = {0732-8818, 1747-1567},
 	url = {https://link.springer.com/10.1007/s40799-024-00715-y},
 	doi = {10.1007/s40799-024-00715-y},
@@ -199,7 +199,7 @@
 }
 
 @article{moser_am_2024,
-	title = {{AM} {Bench} 2022 {Macroscale} {Tensile} {Challenge} at {Different} {Orientations} ({CHAL}-{AMB2022}-04-{MaTTO}) and {Summary} of {Predictions}},
+	title = {{AM} {Bench} 2022: {Macroscale} Tensile Challenge at Different Orientations ({CHAL}-{AMB2022}-04-{MaTTO}) and Summary of Predictions},
 	issn = {2193-9764, 2193-9772},
 	url = {https://link.springer.com/10.1007/s40192-023-00333-3},
 	doi = {10.1007/s40192-023-00333-3},
