### 2025-01-23 67c11d3
Update paper.bib
