+++ paper/paper.bib
@@ -123,7 +123,7 @@
 }
 
 @inbook{3DSlicer,
-	title = {3D Slicer: A Platform for Subject-Specific Image Analysis, Visualization, and Clinical Support},
+	title = {3D {Slicer}: {A} Platform for Subject-Specific Image Analysis, Visualization, and Clinical Support},
 	ISBN = {9781461476573},
 	url = {http://dx.doi.org/10.1007/978-1-4614-7657-3_19},
 	DOI = {10.1007/978-1-4614-7657-3_19},
