### 2025-01-23 #7405 @danielskatz
👋 @NM0ser should the "x" in the title in the Kafka ref be "X"?  If so, please change it.

 Then, please:
