+++ paper/paper.bib
@@ -186,7 +186,7 @@
 }
 
 @article{kafka_insitu_2024,
-	title = {A technique for in-ditu displacement and strain measurement with laboratory-scale x-ray computed tomography},
+	title = {A technique for in-situ displacement and strain measurement with laboratory-scale X-ray computed tomography},
 	issn = {0732-8818, 1747-1567},
 	url = {https://link.springer.com/10.1007/s40799-024-00715-y},
 	doi = {10.1007/s40799-024-00715-y},
