### 2025-01-23 #7405 @NM0ser
Great, I updated the capitalization of the 'X' -- nice catch. Just released version 1.2.0: [IMPPY3D v1.2.0](https://github.com/usnistgov/imppy3d/releases/tag/v1.2.0)

We will next work on getting this release version archived.
