### 2025-01-30 #7405 @danielskatz
👋 @NM0ser - Have you been able to create and deposit the v1.2.0 release?
