### 2025-02-18 #7405 @NM0ser
@danielskatz Perfect timing, I was just about to write an update without actually having seen your message yet. Absolutely. First of all, thanks for your patience. I sent a request to append the existing NIST repository for IMPPY3D with the latest releasing of IMPPY3D (stated above as v1.2.0). While this will not change the DOI of the repository, it does require an internal review. Normally, this is a 1 to 2 weeks turnaround. Things being what they are in the US government right now, there has been some uncertainty and occasional delays throughout NIST. In this particular case, the existing internal reviewer (or data sponsor) had recently left NIST, but the review system did not update. So, apparently, my request to upload an additional file to the repository was routed to a non-existing reviewer unbeknown to me. I decided to check in last Friday, and this is when I learned of what happened. The changes to the repository have now been re-routed to a new person. So, fingers crossed, we are back on track to getting the final release version of IMPPY3D archived in the amended NIST repository. Still not entirely sure of the timeline, but I think a 1 week type of thing for the review to be completed, as originally thought, is a reasonable guess as any. Again, thanks for your patience!
