### 2025-04-01 #7405 @NM0ser
👋 @danielskatz Thanks for checking in. I actually started the repository on Zenodo this morning but got caught up in some meetings today. I have finished a draft of it and created a temporary preview link here. If this looks good to you, I'll publish the archive.

https://zenodo.org/records/15122157?preview=1&token=eyJhbGciOiJIUzUxMiJ9.eyJpZCI6ImI2NThhZmEyLWI2MWItNGE3ZS04MmJmLTI1NGNlNDBkY2FhZiIsImRhdGEiOnt9LCJyYW5kb20iOiI4ZTYyZDZmYTBjOWQ0Zjg3ODQxOTgwN2RjY2QzYTViYSJ9.dRfmfufcNgPPCae5cCM4kczWyUWjEGHjZnd71czTTzewXcxO_OkqGIpTVDpYRp8MLn_l50wKNSHp6ID1JjAXYg
