### 2025-04-01 #7405 @danielskatz
Yes, this looks good to me.
