### 2025-04-02 #7405 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
