### 2025-04-02 #7405 @danielskatz
Congratulations to @NM0ser (Newell Moser) and co-authors on your publication!!

And thanks to @gknapp1 and @sitic for reviewing!
JOSS depends on volunteers, and we couldn't do this without you
